moved present method from Abstract indexable tag presenter to abstract-indexable-presenter
