+++ src/presenters/abstract-indexable-presenter.php
@@ -47,6 +47,27 @@ abstract class Abstract_Indexable_Presenter extends Abstract_Presenter {
 	abstract public function get();
 
 	/**
+	 * Returns a tag in the head.
+	 *
+	 * @return string The tag.
+	 *
+	 * @throws \InvalidArgumentException When a subclass does not define a key property. This should appear during development.
+	 */
+	public function present() {
+		if ( $this->key === 'NO KEY PROVIDED' ) {
+			throw new \InvalidArgumentException( \get_class( $this ) . ' is an Abstract_Indexable_Presenter but does not override the key property.' );
+		}
+
+		$value = $this->get();
+
+		if ( \is_string( $value ) && $value !== '' ) {
+			return \sprintf( $this->tag_format, $this->escape_value( $value ), $this->key );
+		}
+
+		return '';
+	}
+
+	/**
 	 * Transforms an indexable presenter's key to a json safe key string.
 	 *
 	 * @return string
