+++ src/presenters/abstract-indexable-tag-presenter.php
@@ -29,34 +29,13 @@ abstract class Abstract_Indexable_Tag_Presenter extends Abstract_Indexable_Prese
 	protected $escaping = 'attribute';
 
 	/**
-	 * Returns a tag in the head.
-	 *
-	 * @return string The tag.
-	 *
-	 * @throws \InvalidArgumentException When a subclass does not define a key property. This should appear during development.
-	 */
-	public function present() {
-		if ( $this->key === 'NO KEY PROVIDED' ) {
-			throw new \InvalidArgumentException( \get_class( $this ) . ' is an Abstract_Indexable_Tag_Presenter but does not override the key property.' );
-		}
-
-		$value = $this->get();
-
-		if ( \is_string( $value ) && $value !== '' ) {
-			return \sprintf( $this->tag_format, $this->escape( $value ), $this->key );
-		}
-
-		return '';
-	}
-
-	/**
 	 * Escaped the output.
 	 *
 	 * @param string $value The desired method of escaping; 'html', 'url' or 'attribute'.
 	 *
 	 * @return string The escaped value.
 	 */
-	protected function escape( $value ) {
+	protected function escape_value($value ) {
 		switch ( $this->escaping ) {
 			case 'html':
 				return \esc_html( $value );
