+++ tests/unit/presenters/abstract-indexable-presenter-test.php
@@ -40,4 +40,35 @@ class Abstract_Indexable_Presenter_Test extends TestCase {
 
 		$this->assertSame( 'the replaced value', $instance->replace_vars( 'the given value' ) );
 	}
+
+	/**
+	 * Verifies that an exception is thrown, if a subclass does not specify the 'key' property.
+	 *
+	 * @covers \Yoast\WP\SEO\Presenters\Abstract_Indexable_Presenter
+	 */
+	public function test_key_not_defined() {
+		$instance = new Concrete_Indexable_Presenter();
+
+		$this->expectException( 'InvalidArgumentException' );
+		$this->expectExceptionMessage( 'Yoast\WP\SEO\Tests\Unit\Presenters\Concrete_Indexable_Presenter is an Abstract_Indexable_Presenter but does not override the key property.' );
+
+		$instance->present();
+	}
+}
+
+/**
+ * Class Concrete_Indexable_Presenter.
+ *
+ * @phpcs:disable Generic.Files.OneObjectStructurePerFile.MultipleFound
+ */
+class Concrete_Indexable_Presenter extends Abstract_Indexable_Presenter {
+
+	/**
+	 * A concrete implementation of get().
+	 *
+	 * @return string
+	 */
+	public function get() {
+		return 'concrete';
+	}
 }
